back-end/requirements.txt
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
capstone==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colored-traceback==0.3.0
comm==0.1.4
contourpy==1.1.1
cryptography==39.0.2
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==4.2.0
djangorestframework==3.14.0
djfernet==0.8.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.3
fonttools==4.42.1
fqdn==1.5.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
intervaltree==3.1.0
ipykernel==6.25.2
ipython==8.13.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib
matplotlib-inline==0.1.6
mistune==3.0.1
mysql-client==0.0.1
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook_shim==0.2.3
numpy==1.24.3
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plumbum==1.8.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pwntools==4.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyelftools==0.29
Pygments==2.16.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ROPGadget==7.4
rpds-py==0.10.2
rpyc==5.3.1
rsa==4.9
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
unicorn==2.0.1.post1
uri-template==1.3.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.7
word2number==1.1
yachalk==0.1.5
zipp==3.16.2
